\name{print.toswp}
\alias{print.toswp}
\title{Print \code{toswp} class object.
}
\description{Prints introduction to \code{toswp} class object.
}
\usage{
\method{print}{toswp}(x, ...)
}
\arguments{
  \item{x}{The print.toswp object that you wish to print
}
  \item{\dots}{Other arguments to \code{print}
}
}
\details{Just prints out some basic facts about the object.
}
\value{Nothing!
}
\references{
Cardinali, A. and Nason, G.P. (2016) Practical Powerful Wavelet Packet Tests
        for Second-Order Stationarity. Applied and Computational Harmonic
        Analysis, 2016, DOI=10.1016/j.acha.2016.06.006.
}
\author{
G.P.Nason
}

\seealso{\code{\link{WPTOSpickout}}, \code{\link{plot.toswp}},
	\code{\link{summary.toswp}}
}
\examples{
#
# See example in \code{\link{WPTOSpickout}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ts}
